moved how to use content to nav bar and separate component, added tailwindcss typography, and fixed some explanations
