+++ apps/juice/package.json
@@ -25,6 +25,7 @@
     "tailwindcss-animate": "^1.0.7"
   },
   "devDependencies": {
+    "@tailwindcss/typography": "^0.5.19",
     "@tailwindcss/vite": "^4.2.2",
     "chart.js": "^4.5.1",
     "react-chartjs-2": "^5.3.1",
