+++ apps/juice/src/components/battery/AnalysisNav.tsx
@@ -1,4 +1,5 @@
 import { Button } from "@repo/ui/components/button";
+import { Tabs, TabsList, TabsTrigger } from "@repo/ui/components/tabs";
 import { useEffect, useRef, useState } from "react";
 import { JuiceNav } from "../JuiceNav";
 
@@ -41,22 +42,15 @@ export function AnalysisNav({ onNewFiles }: AnalysisNavProps) {
     <div className="sticky top-0 z-50">
       <JuiceNav active="analyzer" />
       <div className="flex items-center gap-4 border-b border-border bg-background/95 px-5 py-1 backdrop-blur-sm">
-        <nav className="flex flex-1 gap-0 overflow-x-auto">
-          {SECTIONS.map(({ id, label }) => (
-            <a
-              key={id}
-              href={`#${id}`}
-              onClick={() => setActiveSection(id)}
-              className={`whitespace-nowrap border-b-2 px-4 py-2 font-mono text-[10px] uppercase tracking-wider transition-colors ${
-                activeSection === id
-                  ? "border-primary text-primary"
-                  : "border-transparent text-muted-foreground hover:text-foreground"
-              }`}
-            >
-              {label}
-            </a>
-          ))}
-        </nav>
+        <Tabs value={activeSection} onValueChange={setActiveSection}>
+          <TabsList>
+            {SECTIONS.map(({ id, label }) => (
+              <TabsTrigger key={id} value={id} asChild>
+                <a href={`#${id}`}>{label}</a>
+              </TabsTrigger>
+            ))}
+          </TabsList>
+        </Tabs>
         <Button size="sm" onClick={onNewFiles}>
           ⬆ New files
         </Button>
