+++ apps/juice/wrangler.jsonc
@@ -1,17 +1,4 @@
 {
   "name": "juice",
-  "main": "@astrojs/cloudflare/entrypoints/server",
-  "compatibility_date": "2026-04-15",
-  "compatibility_flags": ["nodejs_compat"],
-  "kv_namespaces": [
-    {
-      "binding": "SESSION"
-    }
-  ],
-  "images": {
-    "binding": "IMAGES"
-  },
-  "assets": {
-    "binding": "ASSETS"
-  }
+  "compatibility_flags": ["nodejs_compat"]
 }
